Sprint 6: Refactor: extract FastLinkedList and Node classes into separate util package
