+++ src/manager/InMemoryHistoryManager.java
@@ -1,10 +1,8 @@
 package manager;
 
 import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.HashMap;
 
-import model.Node;
+import util.FastLinkedList;
 import model.Task;
 
 public class InMemoryHistoryManager implements HistoryManager {
@@ -26,39 +24,5 @@ public class InMemoryHistoryManager implements HistoryManager {
     @Override
     public ArrayList<Task> getHistory() {
         return history.getTasks();
-    }
-}
-
-class FastLinkedList {
-    private final LinkedList<Node<Task>> tasks = new LinkedList<>();
-    private final HashMap<Integer, Node<Task>> tasksMap = new HashMap<>();
-
-    public void linkLast(Task task) {
-        // existing
-        int id = task.getId();
-        removeTask(id);
-        Node<Task> node = new Node<>(task);
-
-        tasksMap.put(task.getId(), node);
-        tasks.addLast(node);
-    }
-
-    public void removeTask(int id) {
-        if (tasksMap.containsKey(id)) {
-            Node<Task> existingNode = tasksMap.get(id);
-            tasksMap.remove(id);
-
-            tasks.remove(existingNode);
-        }
-    }
-
-    public ArrayList<Task> getTasks() {
-        ArrayList<Task> list = new ArrayList<>();
-
-        for (Node<Task> node : tasks) {
-            list.add(node.data);
-        }
-
-        return list;
     }
 }
