+++ src/util/Node.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 public class Node<T> {
     public T data;
